v4.0 (build 276)

+++ obv_engine/engine/src/main/java/io/olvid/engine/identity/databases/ContactDevice.java
@@ -145,7 +145,7 @@ public class ContactDevice implements ObvDatabase {
             throw new SQLException();
         }
         this.serializedDeviceCapabilities = res.getBytes(SERIALIZED_DEVICE_CAPABILITIES);
-        this.latestChannelCreationPingTimestamp = res.getByte(LATEST_CHANNEL_CREATION_PING_TIMESTAMP);
+        this.latestChannelCreationPingTimestamp = res.getLong(LATEST_CHANNEL_CREATION_PING_TIMESTAMP);
         byte[] preKeyIdBytes = res.getBytes(PRE_KEY_ID);
         this.preKeyId = (preKeyIdBytes == null) ? null : new KeyId(preKeyIdBytes);
         byte[] preKeyEncodedPublicKeyBytes = res.getBytes(PRE_KEY_ENCRYPTION_PUBLIC_KEY);
